Rename middleware files
